refactor: move Header Height constant to constants.ts
