+++ src/components/dashboard/club/common/SideBar.tsx
@@ -4,8 +4,7 @@ import { useEffect, useState } from "react";
 import Link from "next/link";
 import { usePathname, useRouter } from "next/navigation";
 import { cn } from "@/lib/utils";
-import { CLUB_DASHBOARD_ENDPOINT } from "@/lib/constants";
-import { HEADER_HEIGHT } from "@/components/header/Header";
+import { CLUB_DASHBOARD_ENDPOINT, HEADER_HEIGHT } from "@/lib/constants";
 import { ChevronDown } from "@/assets/icons/chevron";
 
 const clubDashboardMenus = [
