+++ src/lib/constants.ts
@@ -1,1 +1,4 @@
+export const HEADER_HEIGHT = "60px";
+
 export const CLUB_DASHBOARD_ENDPOINT = "/dashboard/club";
+export const COMPANY_DASHBOARD_ENDPOINT = "/dashboard/company";
